split healthcheck file
