+++ ydb/core/health_check/self_check_request.h
@@ -1,16 +1,8 @@
+#pragma once
+
+#include "log.h"
 #include "health_check.h"
-
-#include <yql/essentials/public/issue/yql_issue_message.h>
-#include <yql/essentials/public/issue/yql_issue.h>
-
-#include <ydb/library/actors/core/interconnect.h>
-#include <ydb/library/actors/core/hfunc.h>
-#include <ydb/library/actors/interconnect/interconnect.h>
-#include <library/cpp/digest/old_crc/crc.h>
-#include <library/cpp/protobuf/json/proto2json.h>
-#include <ydb/public/sdk/cpp/src/library/grpc/client/grpc_client_low.h>
-
-#include <util/random/shuffle.h>
+#include "self_check_result.h"
 
 #include <ydb/core/base/hive.h>
 #include <ydb/core/base/path.h>
@@ -18,7 +10,6 @@
 #include <ydb/core/base/tablet_pipe.h>
 #include <ydb/core/cms/console/configs_dispatcher.h>
 #include <ydb/core/mon/mon.h>
-#include <ydb/core/base/nameservice.h>
 #include <ydb/core/blobstorage/base/blobstorage_events.h>
 #include <ydb/core/cms/console/console.h>
 #include <ydb/core/mind/tenant_slot_broker.h>
@@ -31,12 +22,11 @@
 #include <ydb/core/protos/blobstorage_distributed_config.pb.h>
 #include <ydb/core/protos/config.pb.h>
 #include <ydb/core/sys_view/common/events.h>
-
-#include <ydb/public/api/grpc/ydb_monitoring_v1.grpc.pb.h>
-#include <regex>
-
+#include <ydb/library/actors/interconnect/interconnect.h>
 #include <ydb/library/actors/wilson/wilson_span.h>
 #include <ydb/library/wilson_ids/wilson.h>
+
+#include <regex>
 
 static decltype(auto) make_vslot_tuple(const NKikimrBlobStorage::TVSlotId& id) {
     return std::make_tuple(id.GetNodeId(), id.GetPDiskId(), id.GetVSlotId());
@@ -132,26 +122,6 @@ public:
     {}
 
     using TGroupId = ui32;
-
-    enum class ETags {
-        None,
-        DBState,
-        StorageState,
-        PoolState,
-        GroupState,
-        VDiskState,
-        PDiskState,
-        NodeState,
-        VDiskSpace,
-        PDiskSpace,
-        ComputeState,
-        TabletState,
-        SystemTabletState,
-        OverloadState,
-        SyncState,
-        Uptime,
-        QuotaUsage,
-    };
 
     enum ETimeoutTag {
         TimeoutBSC,
@@ -271,199 +241,6 @@ public:
         std::vector<const NKikimrSysView::TVSlotEntry*> VSlots;
         ui32 Generation;
         bool LayoutCorrect = true;
-    };
-
-    struct TSelfCheckResult {
-        struct TIssueRecord {
-            Ydb::Monitoring::IssueLog IssueLog;
-            ETags Tag;
-        };
-
-        Ydb::Monitoring::StatusFlag::Status OverallStatus = Ydb::Monitoring::StatusFlag::GREY;
-        TList<TIssueRecord> IssueRecords;
-        Ydb::Monitoring::Location Location;
-        int Level;
-        TString Type;
-        TSelfCheckResult* Upper = nullptr;
-
-        // first level
-        TSelfCheckResult(const TString& type = "")
-            : Level(1)
-            , Type(type)
-            , Upper(nullptr)
-        {}
-
-        TSelfCheckResult(TSelfCheckResult* upper, const TString& type = "")
-            : Type(type)
-            , Upper(upper)
-        {
-            if (Upper) {
-                Location.CopyFrom(upper->Location);
-                Level = upper->Level + 1;
-            }
-        }
-
-        ~TSelfCheckResult() {
-            if (Upper) {
-                Upper->InheritFrom(*this);
-            }
-        }
-
-        static bool IsErrorStatus(Ydb::Monitoring::StatusFlag::Status status) {
-            return status != Ydb::Monitoring::StatusFlag::GREEN;
-        }
-
-        static TString crc16(const TString& data) {
-            return Sprintf("%04x", (ui32)::crc16(data.data(), data.size()));
-        }
-
-        static TString crc32(const TString& data) {
-            return Sprintf("%08x", (ui32)::crc32(data.data(), data.size()));
-        }
-
-        static TString GetIssueId(const Ydb::Monitoring::IssueLog& issueLog) {
-            const Ydb::Monitoring::Location& location(issueLog.location());
-            TStringStream id;
-            if (issueLog.status() != Ydb::Monitoring::StatusFlag::UNSPECIFIED) {
-                id << Ydb::Monitoring::StatusFlag_Status_Name(issueLog.status()) << '-';
-            }
-            id << crc16(issueLog.message());
-            if (location.database().name()) {
-                id << '-' << crc32(location.database().name());
-            }
-            if (location.storage().node().id()) {
-                id << '-' << location.storage().node().id();
-            } else {
-                if (location.storage().node().host()) {
-                    id << '-' << location.storage().node().host();
-                }
-                if (location.storage().node().port()) {
-                    id << '-' << location.storage().node().port();
-                }
-            }
-            if (!location.storage().pool().group().vdisk().id().empty()) {
-                id << '-' << location.storage().pool().group().vdisk().id()[0];
-            } else {
-                if (!location.storage().pool().group().id().empty()) {
-                    id << '-' << location.storage().pool().group().id()[0];
-                } else {
-                    if (location.storage().pool().name()) {
-                        id << '-' << crc32(location.storage().pool().name());
-                    }
-                }
-            }
-            if (!location.storage().pool().group().vdisk().pdisk().empty() && location.storage().pool().group().vdisk().pdisk()[0].id()) {
-                id << '-' << location.storage().pool().group().vdisk().pdisk()[0].id();
-            }
-            if (location.compute().node().id()) {
-                id << '-' << location.compute().node().id();
-            } else {
-                if (location.compute().node().host()) {
-                    id << '-' << location.compute().node().host();
-                }
-                if (location.compute().node().port()) {
-                    id << '-' << location.compute().node().port();
-                }
-            }
-            if (location.compute().pool().name()) {
-                id << '-' << location.compute().pool().name();
-            }
-            if (location.compute().tablet().type()) {
-                id << '-' << location.compute().tablet().type();
-            }
-            if (location.compute().schema().path()) {
-                id << '-' << crc32(location.compute().schema().path());
-            }
-            return id.Str();
-        }
-
-        void ReportStatus(Ydb::Monitoring::StatusFlag::Status status,
-                          const TString& message = {},
-                          ETags setTag = ETags::None,
-                          std::initializer_list<ETags> includeTags = {}) {
-            OverallStatus = MaxStatus(OverallStatus, status);
-            if (IsErrorStatus(status)) {
-                std::vector<TString> reason;
-                if (includeTags.size() != 0) {
-                    for (const TIssueRecord& record : IssueRecords) {
-                        for (const ETags& tag : includeTags) {
-                            if (record.Tag == tag) {
-                                reason.push_back(record.IssueLog.id());
-                                break;
-                            }
-                        }
-                    }
-                }
-                std::sort(reason.begin(), reason.end());
-                reason.erase(std::unique(reason.begin(), reason.end()), reason.end());
-                TIssueRecord& issueRecord(*IssueRecords.emplace(IssueRecords.begin()));
-                Ydb::Monitoring::IssueLog& issueLog(issueRecord.IssueLog);
-                issueLog.set_status(status);
-                issueLog.set_message(message);
-                if (Location.ByteSizeLong() > 0) {
-                    issueLog.mutable_location()->CopyFrom(Location);
-                }
-                issueLog.set_id(GetIssueId(issueLog));
-                if (Type) {
-                    issueLog.set_type(Type);
-                }
-                issueLog.set_level(Level);
-                if (!reason.empty()) {
-                    for (const TString& r : reason) {
-                        issueLog.add_reason(r);
-                    }
-                }
-                if (setTag != ETags::None) {
-                    issueRecord.Tag = setTag;
-                }
-            }
-        }
-
-        bool HasTags(std::initializer_list<ETags> tags) const {
-            for (const TIssueRecord& record : IssueRecords) {
-                for (const ETags tag : tags) {
-                    if (record.Tag == tag) {
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-
-        Ydb::Monitoring::StatusFlag::Status FindMaxStatus(std::initializer_list<ETags> tags) const {
-            Ydb::Monitoring::StatusFlag::Status status = Ydb::Monitoring::StatusFlag::GREY;
-            for (const TIssueRecord& record : IssueRecords) {
-                for (const ETags tag : tags) {
-                    if (record.Tag == tag) {
-                        status = MaxStatus(status, record.IssueLog.status());
-                    }
-                }
-            }
-            return status;
-        }
-
-        void ReportWithMaxChildStatus(const TString& message = {},
-                                        ETags setTag = ETags::None,
-                                        std::initializer_list<ETags> includeTags = {}) {
-            if (HasTags(includeTags)) {
-                ReportStatus(FindMaxStatus(includeTags), message, setTag, includeTags);
-            }
-        }
-
-        Ydb::Monitoring::StatusFlag::Status GetOverallStatus() const {
-            return OverallStatus;
-        }
-
-        void SetOverallStatus(Ydb::Monitoring::StatusFlag::Status status) {
-            OverallStatus = status;
-        }
-
-        void InheritFrom(TSelfCheckResult& lower) {
-            if (lower.GetOverallStatus() >= OverallStatus) {
-                OverallStatus = lower.GetOverallStatus();
-            }
-            IssueRecords.splice(IssueRecords.end(), std::move(lower.IssueRecords));
-        }
     };
 
     template<typename T>
@@ -1738,14 +1515,6 @@ public:
             auto pDiskId = GetPDiskId(pDisk.GetKey());
             PDisksMap.emplace(pDiskId, &pDisk);
         }
-    }
-
-    static Ydb::Monitoring::StatusFlag::Status MaxStatus(Ydb::Monitoring::StatusFlag::Status a, Ydb::Monitoring::StatusFlag::Status b) {
-        return static_cast<Ydb::Monitoring::StatusFlag::Status>(std::max<int>(a, b));
-    }
-
-    static Ydb::Monitoring::StatusFlag::Status MinStatus(Ydb::Monitoring::StatusFlag::Status a, Ydb::Monitoring::StatusFlag::Status b) {
-        return static_cast<Ydb::Monitoring::StatusFlag::Status>(std::min<int>(a, b));
     }
 
     static TString GetNodeLocation(const TEvInterconnect::TNodeInfo& nodeInfo) {
@@ -3236,260 +3005,4 @@ public:
     }
 };
 
-template<typename RequestType>
-class TNodeCheckRequest : public TActorBootstrapped<TNodeCheckRequest<RequestType>> {
-public:
-    using TBase = TActorBootstrapped<TNodeCheckRequest<RequestType>>;
-    using TThis = TNodeCheckRequest<RequestType>;
-
-    static constexpr NKikimrServices::TActivity::EType ActorActivityType() { return NKikimrServices::TActivity::MONITORING_REQUEST; }
-
-    struct TEvPrivate {
-        enum EEv {
-            EvResult = EventSpaceBegin(TEvents::ES_PRIVATE),
-            EvError,
-            EvEnd
-        };
-
-        static_assert(EvEnd < EventSpaceEnd(TEvents::ES_PRIVATE), "expected EvEnd < EventSpaceEnd");
-
-        struct TEvResult : TEventLocal<TEvResult, EvResult> {
-            Ydb::Monitoring::NodeCheckResponse Response;
-
-            TEvResult(Ydb::Monitoring::NodeCheckResponse&& response)
-                : Response(std::move(response))
-            {}
-        };
-
-        struct TEvError : TEventLocal<TEvError, EvError> {
-            NYdbGrpc::TGrpcStatus Status;
-
-            TEvError(NYdbGrpc::TGrpcStatus&& status)
-                : Status(std::move(status))
-            {}
-        };
-    };
-
-    TDuration Timeout = TDuration::MilliSeconds(10000);
-    std::shared_ptr<NYdbGrpc::TGRpcClientLow> GRpcClientLow;
-    TActorId Sender;
-    THolder<RequestType> Request;
-    ui64 Cookie;
-    Ydb::Monitoring::SelfCheckResult Result;
-
-    TNodeCheckRequest(std::shared_ptr<NYdbGrpc::TGRpcClientLow> grpcClient, const TActorId& sender, THolder<RequestType> request, ui64 cookie)
-        : GRpcClientLow(grpcClient)
-        , Sender(sender)
-        , Request(std::move(request))
-        , Cookie(cookie)
-    {
-        Result.set_self_check_result(Ydb::Monitoring::SelfCheck_Result::SelfCheck_Result_UNSPECIFIED);
-    }
-
-    void Bootstrap();
-
-    void AddIssue(Ydb::Monitoring::StatusFlag::Status status, const TString& message) {
-        auto* issue = Result.add_issue_log();
-        issue->set_id(std::to_string(Result.issue_log_size()));
-        issue->set_status(status);
-        issue->set_message(message);
-    }
-
-    void Handle(TEvWhiteboard::TEvSystemStateResponse::TPtr& ev) {
-        NYdbGrpc::TGRpcClientConfig config;
-        for (const auto& systemStateInfo : ev->Get()->Record.GetSystemStateInfo()) {
-            for (const auto& endpoint : systemStateInfo.GetEndpoints()) {
-                if (endpoint.GetName() == "grpc") {
-                    config.Locator = "localhost" + endpoint.GetAddress();
-                    break;
-                } else if (endpoint.GetName() == "grpcs") {
-                    config.Locator = "localhost" + endpoint.GetAddress();
-                    config.EnableSsl = true;
-                    config.SslTargetNameOverride = systemStateInfo.GetHost();
-                    break;
-                }
-            }
-            break;
-        }
-        if (config.Locator.empty()) {
-            AddIssue(Ydb::Monitoring::StatusFlag::RED, "Couldn't find local gRPC endpoint");
-            ReplyAndPassAway();
-        }
-        NActors::TActorSystem* actorSystem = TlsActivationContext->ActorSystem();
-        NActors::TActorId actorId = TBase::SelfId();
-        Ydb::Monitoring::NodeCheckRequest request;
-        NYdbGrpc::TResponseCallback<Ydb::Monitoring::NodeCheckResponse> responseCb =
-            [actorId, actorSystem, context = GRpcClientLow->CreateContext()](NYdbGrpc::TGrpcStatus&& status, Ydb::Monitoring::NodeCheckResponse&& response) -> void {
-            if (status.Ok()) {
-                actorSystem->Send(actorId, new typename TEvPrivate::TEvResult(std::move(response)));
-            } else {
-                actorSystem->Send(actorId, new typename TEvPrivate::TEvError(std::move(status)));
-            }
-        };
-        NYdbGrpc::TCallMeta meta;
-        meta.Timeout = Timeout;
-        auto service = GRpcClientLow->CreateGRpcServiceConnection<::Ydb::Monitoring::V1::MonitoringService>(config);
-        service->DoRequest(request, std::move(responseCb), &Ydb::Monitoring::V1::MonitoringService::Stub::AsyncNodeCheck, meta);
-    }
-
-    void Handle(typename TEvPrivate::TEvResult::TPtr& ev) {
-        auto& operation(ev->Get()->Response.operation());
-        if (operation.ready() && operation.status() == Ydb::StatusIds::SUCCESS) {
-            operation.result().UnpackTo(&Result);
-        } else {
-            Result.set_self_check_result(Ydb::Monitoring::SelfCheck_Result::SelfCheck_Result_MAINTENANCE_REQUIRED);
-            AddIssue(Ydb::Monitoring::StatusFlag::RED, "Local gRPC returned error");
-        }
-        ReplyAndPassAway();
-    }
-
-    void Handle(typename TEvPrivate::TEvError::TPtr& ev) {
-        Result.set_self_check_result(Ydb::Monitoring::SelfCheck_Result::SelfCheck_Result_MAINTENANCE_REQUIRED);
-        AddIssue(Ydb::Monitoring::StatusFlag::RED, "Local gRPC request failed");
-        Y_UNUSED(ev);
-        ReplyAndPassAway();
-    }
-
-    void HandleTimeout() {
-        Result.set_self_check_result(Ydb::Monitoring::SelfCheck_Result::SelfCheck_Result_MAINTENANCE_REQUIRED);
-        AddIssue(Ydb::Monitoring::StatusFlag::RED, "Timeout");
-        ReplyAndPassAway();
-    }
-
-    void StateWork(TAutoPtr<NActors::IEventHandle>& ev) {
-        switch (ev->GetTypeRewrite()) {
-            hFunc(TEvWhiteboard::TEvSystemStateResponse, Handle);
-            hFunc(TEvPrivate::TEvResult, Handle);
-            hFunc(TEvPrivate::TEvError, Handle);
-            cFunc(TEvents::TSystem::Wakeup, HandleTimeout);
-        }
-    }
-
-    void FillResult(Ydb::Monitoring::SelfCheckResult& result) {
-        result = std::move(Result);
-    }
-
-    void ReplyAndPassAway();
-};
-
-template<>
-void TNodeCheckRequest<TEvNodeCheckRequest>::ReplyAndPassAway() {
-    THolder<TEvSelfCheckResult> response = MakeHolder<TEvSelfCheckResult>();
-    Ydb::Monitoring::SelfCheckResult& result = response->Result;
-    FillResult(result);
-    Send(Sender, response.Release(), 0, Cookie);
-    PassAway();
-}
-
-template<>
-void TNodeCheckRequest<NMon::TEvHttpInfo>::ReplyAndPassAway() {
-    static const char HTTPJSON_GOOD[] = "HTTP/1.1 200 Ok\r\nContent-Type: application/json\r\n\r\n";
-    static const char HTTPJSON_NOT_GOOD[] = "HTTP/1.1 500 Failed\r\nContent-Type: application/json\r\n\r\n";
-
-    Ydb::Monitoring::SelfCheckResult result;
-    FillResult(result);
-    auto config = NProtobufJson::TProto2JsonConfig()
-            .SetFormatOutput(false)
-            .SetEnumMode(NProtobufJson::TProto2JsonConfig::EnumName);
-    TStringStream json;
-    if (result.self_check_result() == Ydb::Monitoring::SelfCheck_Result::SelfCheck_Result_GOOD) {
-        json << HTTPJSON_GOOD;
-    } else {
-        json << HTTPJSON_NOT_GOOD;
-    }
-    NProtobufJson::Proto2Json(result, json, config);
-    Send(Sender, new NMon::TEvHttpInfoRes(json.Str(), 0, NMon::IEvHttpInfoRes::EContentType::Custom), 0, Cookie);
-    PassAway();
-}
-
-template<>
-void TNodeCheckRequest<TEvNodeCheckRequest>::Bootstrap() {
-    if (Request->Request.operation_params().has_operation_timeout()) {
-        Timeout = GetDuration(Request->Request.operation_params().operation_timeout());
-    }
-    Result.set_self_check_result(Ydb::Monitoring::SelfCheck_Result::SelfCheck_Result_GOOD);
-    ReplyAndPassAway();
-}
-
-template<>
-void TNodeCheckRequest<NMon::TEvHttpInfo>::Bootstrap() {
-    TActorId whiteboardServiceId = MakeNodeWhiteboardServiceId(TBase::SelfId().NodeId());
-    TBase::Send(whiteboardServiceId, new TEvWhiteboard::TEvSystemStateRequest());
-    const auto& params(Request->Request.GetParams());
-    Timeout = TDuration::MilliSeconds(FromStringWithDefault<ui32>(params.Get("timeout"), Timeout.MilliSeconds()));
-    TBase::Become(&TThis::StateWork, Timeout, new TEvents::TEvWakeup());
-}
-
-class THealthCheckService : public TActorBootstrapped<THealthCheckService> {
-public:
-    static constexpr NKikimrServices::TActivity::EType ActorActivityType() { return NKikimrServices::TActivity::MONITORING_SERVICE; }
-    NKikimrConfig::THealthCheckConfig HealthCheckConfig;
-
-    THealthCheckService()
-    {
-    }
-
-    void Bootstrap() {
-        HealthCheckConfig.CopyFrom(AppData()->HealthCheckConfig);
-        Send(NConsole::MakeConfigsDispatcherID(SelfId().NodeId()),
-             new NConsole::TEvConfigsDispatcher::TEvSetConfigSubscriptionRequest({NKikimrConsole::TConfigItem::HealthCheckConfigItem}));
-        TMon* mon = AppData()->Mon;
-        if (mon) {
-            mon->RegisterActorPage({
-                .RelPath = "status",
-                .ActorSystem = TActivationContext::ActorSystem(),
-                .ActorId = SelfId(),
-                .UseAuth = false,
-            });
-        }
-        Become(&THealthCheckService::StateWork);
-    }
-
-    void Handle(NConsole::TEvConsole::TEvConfigNotificationRequest::TPtr& ev) {
-        const auto& record = ev->Get()->Record;
-        if (record.GetConfig().HasHealthCheckConfig()) {
-            HealthCheckConfig.CopyFrom(record.GetConfig().GetHealthCheckConfig());
-        }
-        Send(ev->Sender, new NConsole::TEvConsole::TEvConfigNotificationResponse(record), 0, ev->Cookie);
-    }
-
-    void Handle(TEvSelfCheckRequest::TPtr& ev) {
-        Register(new TSelfCheckRequest(ev->Sender, ev.Get()->Release(), ev->Cookie, std::move(ev->TraceId), HealthCheckConfig));
-    }
-
-    std::shared_ptr<NYdbGrpc::TGRpcClientLow> GRpcClientLow;
-
-    void Handle(TEvNodeCheckRequest::TPtr& ev) {
-        if (!GRpcClientLow) {
-            GRpcClientLow = std::make_shared<NYdbGrpc::TGRpcClientLow>();
-        }
-        Register(new TNodeCheckRequest<TEvNodeCheckRequest>(GRpcClientLow, ev->Sender, ev.Get()->Release(), ev->Cookie));
-    }
-
-    void Handle(NMon::TEvHttpInfo::TPtr& ev) {
-        if (ev->Get()->Request.GetPath() == "/status") {
-            if (!GRpcClientLow) {
-                GRpcClientLow = std::make_shared<NYdbGrpc::TGRpcClientLow>();
-            }
-            Register(new TNodeCheckRequest<NMon::TEvHttpInfo>(GRpcClientLow, ev->Sender, ev.Get()->Release(), ev->Cookie));
-        } else {
-            Send(ev->Sender, new NMon::TEvHttpInfoRes(NMonitoring::HTTPNOTFOUND, 0, NMon::IEvHttpInfoRes::EContentType::Custom), 0, ev->Cookie);
-        }
-    }
-
-    void StateWork(TAutoPtr<NActors::IEventHandle>& ev) {
-        switch (ev->GetTypeRewrite()) {
-            hFunc(TEvSelfCheckRequest, Handle);
-            hFunc(TEvNodeCheckRequest, Handle);
-            hFunc(NMon::TEvHttpInfo, Handle);
-            hFunc(NConsole::TEvConsole::TEvConfigNotificationRequest, Handle);
-            cFunc(TEvents::TSystem::PoisonPill, PassAway);
-        }
-    }
-};
-
-IActor* CreateHealthCheckService() {
-    return new THealthCheckService();
-}
-
-}
+} // NKikimr::NHealthCheck
