+++ ydb/core/health_check/self_check_result.h
@@ -1,0 +1,230 @@
+#pragma once
+
+#include <ydb/public/api/grpc/ydb_monitoring_v1.grpc.pb.h>
+
+#include <library/cpp/digest/old_crc/crc.h>
+
+namespace NKikimr::NHealthCheck {
+
+    static Ydb::Monitoring::StatusFlag::Status MaxStatus(Ydb::Monitoring::StatusFlag::Status a, Ydb::Monitoring::StatusFlag::Status b) {
+        return static_cast<Ydb::Monitoring::StatusFlag::Status>(std::max<int>(a, b));
+    }
+
+    static Ydb::Monitoring::StatusFlag::Status MinStatus(Ydb::Monitoring::StatusFlag::Status a, Ydb::Monitoring::StatusFlag::Status b) {
+        return static_cast<Ydb::Monitoring::StatusFlag::Status>(std::min<int>(a, b));
+    }
+
+    enum class ETags {
+        None,
+        DBState,
+        StorageState,
+        PoolState,
+        GroupState,
+        VDiskState,
+        PDiskState,
+        NodeState,
+        VDiskSpace,
+        PDiskSpace,
+        ComputeState,
+        TabletState,
+        SystemTabletState,
+        OverloadState,
+        SyncState,
+        Uptime,
+        QuotaUsage,
+    };
+
+    struct TSelfCheckResult {
+        struct TIssueRecord {
+            Ydb::Monitoring::IssueLog IssueLog;
+            ETags Tag;
+        };
+
+        Ydb::Monitoring::StatusFlag::Status OverallStatus = Ydb::Monitoring::StatusFlag::GREY;
+        TList<TIssueRecord> IssueRecords;
+        Ydb::Monitoring::Location Location;
+        int Level;
+        TString Type;
+        TSelfCheckResult* Upper = nullptr;
+
+        // first level
+        TSelfCheckResult(const TString& type = "")
+            : Level(1)
+            , Type(type)
+            , Upper(nullptr)
+        {}
+
+        TSelfCheckResult(TSelfCheckResult* upper, const TString& type = "")
+            : Type(type)
+            , Upper(upper)
+        {
+            if (Upper) {
+                Location.CopyFrom(upper->Location);
+                Level = upper->Level + 1;
+            }
+        }
+
+        ~TSelfCheckResult() {
+            if (Upper) {
+                Upper->InheritFrom(*this);
+            }
+        }
+
+        static bool IsErrorStatus(Ydb::Monitoring::StatusFlag::Status status) {
+            return status != Ydb::Monitoring::StatusFlag::GREEN;
+        }
+
+        static TString crc16(const TString& data) {
+            return Sprintf("%04x", (ui32)::crc16(data.data(), data.size()));
+        }
+
+        static TString crc32(const TString& data) {
+            return Sprintf("%08x", (ui32)::crc32(data.data(), data.size()));
+        }
+
+        static TString GetIssueId(const Ydb::Monitoring::IssueLog& issueLog) {
+            const Ydb::Monitoring::Location& location(issueLog.location());
+            TStringStream id;
+            if (issueLog.status() != Ydb::Monitoring::StatusFlag::UNSPECIFIED) {
+                id << Ydb::Monitoring::StatusFlag_Status_Name(issueLog.status()) << '-';
+            }
+            id << crc16(issueLog.message());
+            if (location.database().name()) {
+                id << '-' << crc32(location.database().name());
+            }
+            if (location.storage().node().id()) {
+                id << '-' << location.storage().node().id();
+            } else {
+                if (location.storage().node().host()) {
+                    id << '-' << location.storage().node().host();
+                }
+                if (location.storage().node().port()) {
+                    id << '-' << location.storage().node().port();
+                }
+            }
+            if (!location.storage().pool().group().vdisk().id().empty()) {
+                id << '-' << location.storage().pool().group().vdisk().id()[0];
+            } else {
+                if (!location.storage().pool().group().id().empty()) {
+                    id << '-' << location.storage().pool().group().id()[0];
+                } else {
+                    if (location.storage().pool().name()) {
+                        id << '-' << crc32(location.storage().pool().name());
+                    }
+                }
+            }
+            if (!location.storage().pool().group().vdisk().pdisk().empty() && location.storage().pool().group().vdisk().pdisk()[0].id()) {
+                id << '-' << location.storage().pool().group().vdisk().pdisk()[0].id();
+            }
+            if (location.compute().node().id()) {
+                id << '-' << location.compute().node().id();
+            } else {
+                if (location.compute().node().host()) {
+                    id << '-' << location.compute().node().host();
+                }
+                if (location.compute().node().port()) {
+                    id << '-' << location.compute().node().port();
+                }
+            }
+            if (location.compute().pool().name()) {
+                id << '-' << location.compute().pool().name();
+            }
+            if (location.compute().tablet().type()) {
+                id << '-' << location.compute().tablet().type();
+            }
+            if (location.compute().schema().path()) {
+                id << '-' << crc32(location.compute().schema().path());
+            }
+            return id.Str();
+        }
+
+        void ReportStatus(Ydb::Monitoring::StatusFlag::Status status,
+                          const TString& message = {},
+                          ETags setTag = ETags::None,
+                          std::initializer_list<ETags> includeTags = {}) {
+            OverallStatus = MaxStatus(OverallStatus, status);
+            if (IsErrorStatus(status)) {
+                std::vector<TString> reason;
+                if (includeTags.size() != 0) {
+                    for (const TIssueRecord& record : IssueRecords) {
+                        for (const ETags& tag : includeTags) {
+                            if (record.Tag == tag) {
+                                reason.push_back(record.IssueLog.id());
+                                break;
+                            }
+                        }
+                    }
+                }
+                std::sort(reason.begin(), reason.end());
+                reason.erase(std::unique(reason.begin(), reason.end()), reason.end());
+                TIssueRecord& issueRecord(*IssueRecords.emplace(IssueRecords.begin()));
+                Ydb::Monitoring::IssueLog& issueLog(issueRecord.IssueLog);
+                issueLog.set_status(status);
+                issueLog.set_message(message);
+                if (Location.ByteSizeLong() > 0) {
+                    issueLog.mutable_location()->CopyFrom(Location);
+                }
+                issueLog.set_id(GetIssueId(issueLog));
+                if (Type) {
+                    issueLog.set_type(Type);
+                }
+                issueLog.set_level(Level);
+                if (!reason.empty()) {
+                    for (const TString& r : reason) {
+                        issueLog.add_reason(r);
+                    }
+                }
+                if (setTag != ETags::None) {
+                    issueRecord.Tag = setTag;
+                }
+            }
+        }
+
+        bool HasTags(std::initializer_list<ETags> tags) const {
+            for (const TIssueRecord& record : IssueRecords) {
+                for (const ETags tag : tags) {
+                    if (record.Tag == tag) {
+                        return true;
+                    }
+                }
+            }
+            return false;
+        }
+
+        Ydb::Monitoring::StatusFlag::Status FindMaxStatus(std::initializer_list<ETags> tags) const {
+            Ydb::Monitoring::StatusFlag::Status status = Ydb::Monitoring::StatusFlag::GREY;
+            for (const TIssueRecord& record : IssueRecords) {
+                for (const ETags tag : tags) {
+                    if (record.Tag == tag) {
+                        status = MaxStatus(status, record.IssueLog.status());
+                    }
+                }
+            }
+            return status;
+        }
+
+        void ReportWithMaxChildStatus(const TString& message = {},
+                                        ETags setTag = ETags::None,
+                                        std::initializer_list<ETags> includeTags = {}) {
+            if (HasTags(includeTags)) {
+                ReportStatus(FindMaxStatus(includeTags), message, setTag, includeTags);
+            }
+        }
+
+        Ydb::Monitoring::StatusFlag::Status GetOverallStatus() const {
+            return OverallStatus;
+        }
+
+        void SetOverallStatus(Ydb::Monitoring::StatusFlag::Status status) {
+            OverallStatus = status;
+        }
+
+        void InheritFrom(TSelfCheckResult& lower) {
+            if (lower.GetOverallStatus() >= OverallStatus) {
+                OverallStatus = lower.GetOverallStatus();
+            }
+            IssueRecords.splice(IssueRecords.end(), std::move(lower.IssueRecords));
+        }
+    };
+
+} // NKikimr::NHealthCheck
